{{cookiecutter.repo_name}}/config/wsgi.py
# -*- coding: utf-8 -*-
"""
WSGI config for {{ cookiecutter.repo_name }} project.

It exposes the WSGI callable as a module-level variable named ``application``.

https://docs.djangoproject.com/en/1.9/howto/deployment/wsgi/
"""

import os

from django.core.wsgi import get_wsgi_application

os.environ.setdefault("DJANGO_SETTINGS_MODULE", "{{ cookiecutter.django_default_settings }}")

application = get_wsgi_application()
