{{cookiecutter.repo_name}}/requirements/dev.txt
# Dependencies for the dev environment
-r base.txt

# test tools
django-extensions>=1.6.1
django-nose==1.4.2
coverage>=4.0.3

# debug tools
django-debug-toolbar>=1.4
